Add indicators for volume and brightness (#170)
* fix: clippy lints in status_bar and clock

* feat: volume and brightness OSD indicators

* refactor: reuse loaded stylesheet in theme settings

* fix: clippy lints in launcher app view

* fix: keep stale key releases in evdev poll

* refactor: reuse ui frame interval for label scrolling

* fix: forward debug-ui feature to common in allium-menu

* fix: resolve app icon path against the pak directory

* fix: stop redrawing an image whose load failed

* perf: read only the plate area when showing the osd

+++ crates/common/src/view/status_bar.rs
@@ -85,8 +85,8 @@ where
         if self.should_draw() && styles.status_bar.status_backdrop_color.a() > 0 {
             let mut bbox = self.row.bounding_box(styles);
             if bbox.w > 0 && bbox.h > 0 {
-                let padding_x = styles.ui.padding_x as i32;
-                let padding_y = styles.ui.padding_y as i32;
+                let padding_x = styles.ui.padding_x;
+                let padding_y = styles.ui.padding_y;
                 bbox.x -= padding_x;
                 bbox.y -= padding_y;
                 bbox.w += (padding_x * 2) as u32;
